Remove apicast-cloud-hosted unavailable policies
It is necessary to not use a runtime image as the deleted policies will be
deleted from the builder but still present in the runtime image. The only way to delete
them is to use directly the builder image. In this case it is not much of a problem as
the final size of the image is the same whether runtime image is used or not.