Add logic to mapping-service to make apicast work in preview
The preview environment uses an obfuscated production dataset. When the mapping-service
requests the admin_domain for the given apicast request in order to get the apicast configs from
that url, system returns the production url for the admin_domain and so subsequent requests
are sent to production.

This commit adds logic so that, based on the existence of the PREVIEW_BASE_DOMAIN environment variable,
mapping-service changes the admin_domain returned by system and points it to the preview environment.