Move build to productized image
A bunch of things here:

1) We move to a single Dockerfile and we move out from s2i; the main reason is
the productized image is not using s2i.

2) LUA: We need to install net-url, and the main reason is that it's not part of
gateway-rockspecs, I need to add that one into the build of the package.

3) LUA: I'm going to add lua-rover too, but now it is installed using luarocks.

4) Perl: Some Test::Nginx changed, so we need to pin the version on the
cpanfile, the reason of `<=` is because if you set a version 0.24 and there is a
newest version, it'll get the more recent.

5) Makefile changes: now should be easy to understand.

Fix THREESCALE-6622

Signed-off-by: Eloy Coto <eloy.coto@acalustra.com>